Return incomplete batches of sliding window sequences.
The sequence might not neatly dividing into a number of sliding window
slices that is a multiple of the batch size. Because of this the last
frames in the sequence might be dropped. In order to minimize the number
of dropped frames the last batch of 'sliding windows' might contain
a number of 'sliding windows' <= given batch_size.