Add Traffic4CastSample transformation and fix minor issues.
Add callables that can be passed as transforms to the Traffic4CastSet to
perform transformations as selecting channels and permuting of axes.

Fix valid frames mask initialization.

Fix number of batch computation for the selected_temporal_batches.

Fix computation of the range of frames that can be used in random batch
generation.