Improve model type selection for the train application.
Change the --city argument to --cities in order to allow for training on
multiple cities.
Move --model argument to --model-type and have it designate a type of
model.
Use the --model argument to designate the path to a model of type
--model-type to load and finetune train it.
Add default naming for models that are trained from scratch. Default
name is a concatenation of model type, channels and cities.